% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_generic.R
\name{process_generic}
\alias{process_generic}
\title{Process batch generic data}
\usage{
process_generic(
  multiassay,
  scale_data = TRUE,
  remove_sample_outliers = FALSE,
  process_func = function(x) return(x)
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{process_func}{}
}
\value{
a MultiAssayExperiment object with \code{generic_processed} slot
}
\description{
Process batch generic data
}
\seealso{
Other Pre-processing: 
\code{\link{batch_correction_protein}()},
\code{\link{combat_correction}()},
\code{\link{filter_protein}()},
\code{\link{generate_multiassay}()},
\code{\link{median_centering_correction}()},
\code{\link{process_protein}()},
\code{\link{process_rna}()}
}
\concept{Pre-processing}
